ft_printf_1.0

+++ ft_printf.h
@@ -18,13 +18,14 @@
 # include <stdarg.h>
 # include <stdio.h>
 # include <limits.h>
+# include <stdint.h>
 
 int	ft_printf(const char *format, ...);
 int	ft_putchar(int c);
 int	ft_putstr(char *s);
 int	ft_putnbr(int n);
 int	ft_putnbr_uns(unsigned int n);
-int	ft_puthex(unsigned int n);
+int	ft_puthex(uintptr_t n);
 int	ft_puthex_up(unsigned int n);
 int	ft_putptr(void *ptr);
 #endif
